Update go.mod and docker to account for api->rest rename
