+++ rest/Makefile
@@ -3,7 +3,7 @@ EXEC_NAME?=go-api
 DOCKER_RUNNER?=docker
 REGISTRY?=localhost:5000
 RELEASE_TAG=$(shell git rev-parse --short HEAD)
-BASE_TAG=$(REGISTRY)/utdnebula/api/go-api
+BASE_TAG=$(REGISTRY)/utdnebula/rest/go-api
 
 setup:
 	go install honnef.co/go/tools/cmd/staticcheck@latest
@@ -28,6 +28,6 @@ build: server.go
 clean: $(EXEC_NAME)
 	rm $(EXEC_NAME)
 
-docker: Dockerfile common/ configs/ controllers/ docs/ responses/ routes/ schema/ server.go
+docker: Dockerfile configs/ controllers/ docs/ responses/ routes/ schema/ server.go
 	$(DOCKER_RUNNER) build -t $(BASE_TAG):$(RELEASE_TAG) .
 	$(DOCKER_RUNNER) tag $(BASE_TAG):$(RELEASE_TAG) $(BASE_TAG):latest
