+++ rest/controllers/club.go
@@ -8,8 +8,8 @@ import (
 
 	"encoding/json"
 
-	"github.com/UTDNebula/nebula-api/api/configs"
-	"github.com/UTDNebula/nebula-api/api/schema"
+	"github.com/UTDNebula/nebula-api/rest/configs"
+	"github.com/UTDNebula/nebula-api/rest/schema"
 	"github.com/gin-gonic/gin"
 )
