+++ rest/controllers/email.go
@@ -9,7 +9,7 @@ import (
 	"os"
 	"time"
 
-	"github.com/UTDNebula/nebula-api/api/schema"
+	"github.com/UTDNebula/nebula-api/rest/schema"
 	"github.com/gin-gonic/gin"
 	"github.com/wneessen/go-mail"
