+++ rest/controllers/evaluation.go
@@ -13,9 +13,9 @@ import (
 	"time"
 
 	"log"
-	"github.com/UTDNebula/nebula-api/api/configs"
+	"github.com/UTDNebula/nebula-api/rest/configs"
 
-	"github.com/UTDNebula/nebula-api/api/schema"
+	"github.com/UTDNebula/nebula-api/rest/schema"
 
 	"github.com/gin-gonic/gin"
